+++ lib/hexapdf/cli.rb
@@ -38,6 +38,7 @@ require 'hexapdf/cli/inspect'
 require 'hexapdf/cli/modify'
 require 'hexapdf/cli/merge'
 require 'hexapdf/cli/optimize'
+require 'hexapdf/cli/images'
 require 'hexapdf/version'
 require 'hexapdf/document'
 
@@ -68,6 +69,7 @@ module HexaPDF
         main_command.options.version = HexaPDF::VERSION
         add_command(HexaPDF::CLI::Info.new)
         add_command(HexaPDF::CLI::Files.new)
+        add_command(HexaPDF::CLI::Images.new)
         add_command(HexaPDF::CLI::Inspect.new)
         add_command(HexaPDF::CLI::Modify.new)
         add_command(HexaPDF::CLI::Optimize.new)